+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - (Feature) (Platform) MetaV1 GetBatch
 - (Feature) Switch Endpoints to EndpointSlices
 - (Feature) Ensure License Grace Period
+- (Feature) (Platform) Allow Service Chart change
 
 ## [1.4.1](https://github.com/arangodb/kube-arangodb/tree/1.4.1) (2026-01-27)
 - (Bugfix) (Platform) Ensure Helm History limit in ArangoPlatformService